+++ src/test/java/com/jnape/palatable/lambda/lens/lenses/MapLensTest.java
@@ -6,6 +6,7 @@ import org.junit.Test;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.LinkedHashMap;
 import java.util.Map;
 
 import static com.jnape.palatable.lambda.adt.Maybe.just;
@@ -15,7 +16,6 @@ import static com.jnape.palatable.lambda.lens.functions.Set.set;
 import static com.jnape.palatable.lambda.lens.functions.View.view;
 import static com.jnape.palatable.lambda.lens.lenses.MapLens.keys;
 import static com.jnape.palatable.lambda.lens.lenses.MapLens.mappingValues;
-import static com.jnape.palatable.lambda.lens.lenses.MapLens.valueAt;
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyMap;
 import static java.util.Collections.emptySet;
@@ -27,11 +27,12 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotSame;
 import static org.junit.Assert.assertThat;
 import static testsupport.assertion.LensAssert.assertLensLawfulness;
+import static testsupport.matchers.IterableMatcher.iterates;
 
 public class MapLensTest {
 
     @Test
-    public void asCopyFocusesOnMapThroughCopy() {
+    public void asCopy() {
         assertLensLawfulness(MapLens.asCopy(),
                              asList(emptyMap(), singletonMap("foo", 1), new HashMap<String, Integer>() {{
                                  put("foo", 1);
@@ -46,8 +47,29 @@ public class MapLensTest {
     }
 
     @Test
-    public void valueAtFocusesOnValueAtKey() {
-        assertLensLawfulness(valueAt("foo"),
+    public void asCopyWithCopyFn() {
+        assertLensLawfulness(MapLens.asCopy(LinkedHashMap::new),
+                             asList(emptyMap(), singletonMap("foo", 1), new HashMap<String, Integer>() {{
+                                 put("foo", 1);
+                                 put("bar", 2);
+                                 put("baz", 3);
+                             }}),
+                             asList(emptyMap(), singletonMap("foo", 1), new HashMap<String, Integer>() {{
+                                 put("foo", 1);
+                                 put("bar", 2);
+                                 put("baz", 3);
+                             }}));
+
+        assertThat(view(MapLens.asCopy(LinkedHashMap::new), new LinkedHashMap<String, Integer>() {{
+            put("foo", 1);
+            put("bar", 2);
+            put("baz", 3);
+        }}).keySet(), iterates("foo", "bar", "baz"));
+    }
+
+    @Test
+    public void valueAt() {
+        assertLensLawfulness(MapLens.valueAt("foo"),
                              asList(emptyMap(), singletonMap("foo", 1), new HashMap<String, Integer>() {{
                                  put("foo", 1);
                                  put("bar", 2);
@@ -57,8 +79,20 @@ public class MapLensTest {
     }
 
     @Test
-    public void valueAtWithDefaultValueFocusedOnValueAtKey() {
-        Lens.Simple<Map<String, Integer>, Integer> atFoo = valueAt("foo", -1);
+    public void valueAtWithCopyFn() {
+        assertLensLawfulness(MapLens.valueAt("foo"),
+                             asList(emptyMap(), singletonMap("foo", 1), new HashMap<String, Integer>() {{
+                                 put("foo", 1);
+                                 put("bar", 2);
+                                 put("baz", 3);
+                             }}),
+                             asList(nothing(), just(1)));
+    }
+
+
+    @Test
+    public void valueAtWithDefaultValue() {
+        Lens.Simple<Map<String, Integer>, Integer> atFoo = MapLens.valueAt("foo", -1);
 
         assertEquals((Integer) 1, view(atFoo, new HashMap<String, Integer>() {{
             put("foo", 1);